requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
g4f[all]==0.3.3.2
deep-translator==1.11.4
pydub
SpeechRecognition
ffmpeg
langchain-community
langchain
pymongo==4.7.1
Flask==3.0.3
safoneapi==1.0.63
psutil==5.9.8
telegraph==2.2.0
python-docx==1.1.2
python-pptx
PyMuPDF==1.24.5
requests~=2.32.3
TgCrypto
beautifulsoup4~=4.12.3
modelbit~=0.41.19
SpeechRecognition==3.10.4
pydub==0.25.1
pytubefix==8.1.1